requirements.txt
mysql-connector==2.2.9
psycopg2-binary==2.8.3
PyYAML==5.1
requests==2.21.0